feat: introduced linux/windows layer - no changes on the Linux behavior - common modules used across the repository - enablement of platform specific code
