--- src/sys/sys.rs
@@ -1,5 +1,0 @@
-#[cfg(linux)]
-mod linux;
-
-#[cfg(windows)]
-mod windows;
